helper script updated

+++ build_package.sh
@@ -8,17 +8,21 @@ archive_name="libreelec_argondevice_${addon_version}.zip"
 archive_dir=~/Documents/Argon40
 package_dir=${addon_id}
 
-mkdir -p ${package_dir}
-cp changelog.txt ${package_dir}
-cp README.md ${package_dir}
-cp source/addon.xml ${package_dir}
-cp source/default.py ${package_dir}
-cp source/main.py ${package_dir}
-cp source/LICENSE ${package_dir}
-cp -r source/resources ${package_dir}
-if [ -f ${archive_dir}/${archive_name} ]; then
-    echo "Add-on archive for this version already exists."
+if [ -d ${archive_dir} ]; then
+    mkdir -p ${package_dir}
+    cp changelog.txt ${package_dir}
+    cp README.md ${package_dir}
+    cp source/addon.xml ${package_dir}
+    cp source/default.py ${package_dir}
+    cp source/main.py ${package_dir}
+    cp source/LICENSE ${package_dir}
+    cp -r source/resources ${package_dir}
+    if [ -f ${archive_dir}/${archive_name} ]; then
+        echo "Add-on archive for this version already exists."
+    else
+        zip -r9 ${archive_dir}/${archive_name} ${package_dir}
+    fi
+    rm -r ${package_dir}
 else
-    zip -r9 ${archive_dir}/${archive_name} ${package_dir}
+    echo "The archive directory: \"$archive_dir\" doesn't exist. Please adapt the path in this script to your requirements or create the directory beforehand."
 fi
-rm -r ${package_dir}
